Code working, need to add tests to translation manager and look for other possible cleanups.
