+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialVirtualWorkerTranslator.h
@@ -4,7 +4,6 @@
 
 #include "CoreMinimal.h"
 
-#include "Containers/Queue.h"
 #include "SpatialCommonTypes.h"
 #include "SpatialConstants.h"
 
@@ -16,7 +15,6 @@ DECLARE_LOG_CATEGORY_EXTERN(LogSpatialVirtualWorkerTranslator, Log, All)
 class UAbstractLBStrategy;
 class USpatialStaticComponentView;
 class USpatialReceiver;
-class USpatialWorkerConnection;
 
 class SPATIALGDK_API SpatialVirtualWorkerTranslator
 {
@@ -25,15 +23,12 @@ public:
 
 	void Init(UAbstractLBStrategy* InLoadBalanceStrategy,
 		USpatialStaticComponentView* InStaticComponentView,
-		USpatialReceiver* InReceiver,
-		USpatialWorkerConnection* InConnection,
 		PhysicalWorkerName InPhysicalWorkerName);
 
 	// Returns true if the Translator has received the information needed to map virtual workers to physical workers.
 	// Currently that is only the number of virtual workers desired.
 	bool IsReady() const { return bIsReady; }
 
-	void AddVirtualWorkerIds(const TSet<VirtualWorkerId>& InVirtualWorkerIds);
 	VirtualWorkerId GetLocalVirtualWorkerId() const { return LocalVirtualWorkerId; }
 	PhysicalWorkerName GetLocalPhysicalWorkerName() const { return LocalPhysicalWorkerName; }
 
@@ -46,22 +41,11 @@ public:
 	// On receiving a version of the translation state, apply that to the internal mapping.
 	void ApplyVirtualWorkerManagerData(Schema_Object* ComponentObject);
 
-	// Authority may change on one of two components we care about:
-	// 1) The translation component, in which case this worker is now authoritative on the virtual to physical worker translation.
-	// 2) The ACL component for some entity, in which case this worker is now authoritative for the entity and will be
-	// responsible for updating the ACL in the future if this worker loses authority.
-	void AuthorityChanged(const Worker_AuthorityChangeOp& AuthChangeOp);
-
 private:
 	TWeakObjectPtr<UAbstractLBStrategy> LoadBalanceStrategy;
 	TWeakObjectPtr<USpatialStaticComponentView> StaticComponentView;
-	TWeakObjectPtr<USpatialReceiver> Receiver;
-	TWeakObjectPtr<USpatialWorkerConnection> Connection;
 
 	TMap<VirtualWorkerId, PhysicalWorkerName> VirtualToPhysicalWorkerMapping;
-	TQueue<VirtualWorkerId> UnassignedVirtualWorkers;
-
-	bool bWorkerEntityQueryInFlight;
 
 	bool bIsReady;
 
@@ -71,17 +55,7 @@ private:
 
 	// Serialization and deserialization of the mapping.
 	void ApplyMappingFromSchema(Schema_Object* Object);
-	void WriteMappingToSchema(Schema_Object* Object) const;
 	bool IsValidMapping(Schema_Object* Object) const;
-
-	// The following methods are used to query the Runtime for all worker entities and update the mapping
-	// based on the response.
-	void QueryForWorkerEntities();
-	void WorkerEntityQueryDelegate(const Worker_EntityQueryResponseOp& Op);
-	void ConstructVirtualWorkerMappingFromQueryResponse(const Worker_EntityQueryResponseOp& Op);
-	void SendVirtualWorkerMappingUpdate();
-
-	void AssignWorker(const PhysicalWorkerName& WorkerId);
 
 	void UpdateMapping(VirtualWorkerId Id, PhysicalWorkerName Name);
 };
